crypto_portfolio/requirements.txt
anthropic>=0.40.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
torch>=2.0.0
transformers>=4.35.0
schedule>=1.2.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
matplotlib>=3.7.0
seaborn>=0.12.0
pytrends>=4.9.2
tweepy>=4.14.0
praw>=7.7.0
pycoingecko>=3.1.0
python-binance>=1.0.19
pytest>=7.4.0
